Graphical TFT fixes, cleanup (#20861)
